add: PersonBuilder 추가 및 적용
